+++ modules/Scheduling/includes/Courses.fnc.php
@@ -230,8 +230,8 @@ function CoursePeriodOptionInputs( $course_period_RET, $array, $new )
 			(float) $course_period_RET['CREDITS'],
 		$array . '[CREDITS]',
 		_( 'Credits' ),
-		// Fix #329 SQL error division by zero in t_update_mp_stats(): set min Credits to 1.
-		' type="number" step="0.01" min="1" max="9999"',
+		// @since 11.1 SQL set min Credits to 0 & fix division by zero error
+		' type="number" step="0.01" min="0" max="9999"',
 		( ! isset( $course_period_RET['CREDITS'] ) || is_null( $course_period_RET['CREDITS'] ) ? false : true )
 	);